refactor: move retrieval to options.js
