+++ lib/common/options.js
@@ -4,6 +4,7 @@ import fs from 'fs'
 import _ from 'lodash'
 import consola from 'consola'
 
+import compressionMiddleware from 'compression'
 import { isPureObject, isUrl } from '../common/utils'
 
 import modes from './modes'
@@ -169,6 +170,10 @@ Options.from = function (_options) {
     options.ignore.push(`**/${options.ignorePrefix}*.*`)
   }
 
+  // Compression middleware
+
+  options.compressor = retrieveCompressionMiddleware(options.render)
+
   // Apply mode preset
   const modePreset = modes[options.mode || 'universal'] || modes.universal
   _.defaultsDeep(options, modePreset)
@@ -208,4 +213,19 @@ Options.from = function (_options) {
   }
 
   return options
+}
+
+const retrieveCompressionMiddleware = ({ compressor, gzip }) => {
+  if (!compressor) {
+    if (gzip) {
+      consola.warn('build.render.gzip is deprecated and will be removed in a future version! Please switch to' +
+        ' build.render.compressor')
+      return compressionMiddleware(gzip)
+    }
+    return compressionMiddleware()
+  }
+  if (typeof compressor === 'object') {
+    return compressionMiddleware(compressor)
+  }
+  return compressor
 }
