+++ lib/core/renderer.js
@@ -3,7 +3,6 @@ import crypto from 'crypto'
 
 import serialize from 'serialize-javascript'
 import serveStatic from 'serve-static'
-import compressionMiddleware from 'compression'
 import _ from 'lodash'
 import fs from 'fs-extra'
 import { createBundleRenderer } from 'vue-server-renderer'
@@ -203,7 +202,7 @@ export default class Renderer {
 
     // Compression middleware for production
     if (!this.options.dev) {
-      this.useMiddleware(retrieveCompressionMiddleware(this.options.render))
+      this.useMiddleware(this.options.render.compressor)
     }
 
     // Add webpack middleware only for development
@@ -472,18 +471,3 @@ export const resourceMap = [
     transform: parseTemplate
   }
 ]
-
-const retrieveCompressionMiddleware = ({ compressor, gzip }) => {
-  if (!compressor) {
-    if (gzip) {
-      consola.warn('build.render.gzip is deprecated and will be removed in a future version! Please switch to' +
-        ' build.render.compressor')
-      return compressionMiddleware(gzip)
-    }
-    return compressionMiddleware()
-  }
-  if (typeof compressor === 'object') {
-    return compressionMiddleware(compressor)
-  }
-  return compressor
-}
